server/requirements.txt
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
et-xmlfile==1.1.0
exceptiongroup==1.1.1
firebase-admin==6.1.0
Flask==2.3.2
Flask-Cors==3.0.10
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.3
oauth2client==4.1.3
openpyxl==3.1.2
outcome==1.2.0
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyparsing==3.0.9
Pyrebase4==4.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
wsproto==1.2.0
